requirements.txt
groq
faiss-cpu
sentence-transformers
python-dotenv